Changed the structure for the Dockerfile, export the code that needs to be hidden
