+++ Bot/Dockerfile
@@ -1,0 +1,7 @@
+# Build competitor bot image by extending the base
+FROM bot_base:latest
+
+# Example: allow teams to add dependencies
+# RUN pip install --no-cache-dir <extra-dependency>
+
+COPY bot_class.py .
