+++ Detector/Dockerfile
@@ -1,0 +1,7 @@
+# Build competitor detector image by extending the base
+FROM detector_base:latest
+
+# Example: allow teams to add dependencies
+# RUN pip install --no-cache-dir <extra-dependency>
+
+COPY detector_class.py .
