+++ Detector/detector_class.py
@@ -1,0 +1,10 @@
+class Detector:
+    def detect_bot(self, session_data):
+        # todo logic
+        # Example:
+        marked_account = []
+
+        for user in session_data["users"]:
+            marked_account.append({"user_id": f"{user['user_id']}", "confidence": 50, "bot": False})
+
+        return marked_account
