--- run.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-# Run script
-python3 BotTemplate/main_bot.py 
-#python3 DetectorTemplate/main_detector.py
-
-python3 email_results.py
